Coursework V0.8
This version has nearly all the features that are on the main roadmap for the code to have. There is still room for expansion, but this is a nearly out of beta product.

Going to be trying out APIs or similar things to check out how I can remotly access this code.